Basic tests for the new Tree (composite) environment
